pyproject.toml
[project]
name = "llm-router-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.57.1",
    "dspy-ai>=2.6.27",
    "email-validator>=2.2.0",
    "gradio>=5.38.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
